[SPARK-46504][PS][TESTS][FOLLOWUP] Break the remaining part of `IndexesTests` into small test files
### What changes were proposed in this pull request?
Break the remaining part of `IndexesTests` into small test files

### Why are the changes needed?
testing parallelism

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44588 from zhengruifeng/ps_test_idx_base_lastlast.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
