+++ dev/sparktestsupport/modules.py
@@ -795,7 +795,9 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.generic",
         "pyspark.pandas.series",
         # unittests
-        "pyspark.pandas.tests.indexes.test_base",
+        "pyspark.pandas.tests.indexes.test_basic",
+        "pyspark.pandas.tests.indexes.test_getattr",
+        "pyspark.pandas.tests.indexes.test_name",
         "pyspark.pandas.tests.indexes.test_conversion",
         "pyspark.pandas.tests.indexes.test_drop",
         "pyspark.pandas.tests.indexes.test_level",
@@ -1095,7 +1097,9 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.test_parity_sql",
         "pyspark.pandas.tests.connect.test_parity_typedef",
         "pyspark.pandas.tests.connect.test_parity_utils",
-        "pyspark.pandas.tests.connect.indexes.test_parity_base",
+        "pyspark.pandas.tests.connect.indexes.test_parity_basic",
+        "pyspark.pandas.tests.connect.indexes.test_parity_getattr",
+        "pyspark.pandas.tests.connect.indexes.test_parity_name",
         "pyspark.pandas.tests.connect.indexes.test_parity_conversion",
         "pyspark.pandas.tests.connect.indexes.test_parity_drop",
         "pyspark.pandas.tests.connect.indexes.test_parity_level",
